+++ triplea/service/repository/state/custom/affiliation_mining_titipata_integration.py
@@ -96,6 +96,10 @@ def affiliation_mining_titipata_integration(article: Article):
         aff_list = _get_affiliation_integrated_from_wos_scopus(article)
     elif article.SourceBank == SourceBankType.SCOPUS:
         aff_list = _get_affiliation_integrated_from_wos_scopus(article)
+    elif article.SourceBank == SourceBankType.IEEE:
+        aff_list = _get_affiliation_integrated_from_wos_scopus(article)
+    elif article.SourceBank == SourceBankType.UNKNOWN:
+        aff_list = _get_affiliation_integrated_from_wos_scopus(article)
     else:
         raise NotImplementedError